Refactor GET providers
Jira: MODKBEKBJ-109
Changed getEholdingsProviders to use RMAPITemplate
